[#9] JuiceMaker.swift에서 Fruit Manager를 FruitManager.swift로 이동
